+++ src/main/java/hu/oe/nik/szfmv/automatedcar/engine/TractionForce.java
@@ -1,4 +1,4 @@
-package hu.oe.nik.szfmv.automatedcar;
+package hu.oe.nik.szfmv.automatedcar.engine;
 
 public class TractionForce {
